Calculated total price deducting the offers.

+++ ShoppingCart/src/main/java/com/self/test/CheckoutService.java
@@ -1,11 +1,12 @@
 package com.self.test;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 
 import com.self.test.product.Item;
 
 public interface CheckoutService { 
 
-	Double getTotalPrice(Collection<Item> items);
+	BigDecimal getTotalPrice(Collection<Item> items);
 
 }
